Make ZFS 0.8 PPA/module installations conditional to the distro

+++ install-zfs.sh
@@ -360,16 +360,18 @@ function ask_pool_tweaks {
 function install_zfs_module {
   print_step_info_header
 
-  if [[ ${ZFS_SKIP_LIVE_ZFS_MODULE_INSTALL:-} == "" ]]; then
-    echo "zfs-dkms zfs-dkms/note-incompatible-licenses note true" | debconf-set-selections
+  if [[ "${c_linux_setups_zfs_0_8_support["$v_linux_distribution"]}" == "1" ]]; then
+    if [[ ${ZFS_SKIP_LIVE_ZFS_MODULE_INSTALL:-} == "" ]]; then
+      echo "zfs-dkms zfs-dkms/note-incompatible-licenses note true" | debconf-set-selections
 
-    add-apt-repository --yes ppa:jonathonf/zfs
-    apt install --yes zfs-dkms
+      add-apt-repository --yes ppa:jonathonf/zfs
+      apt install --yes zfs-dkms
 
-    systemctl stop zfs-zed
-    modprobe -r zfs
-    modprobe zfs
-    systemctl start zfs-zed
+      systemctl stop zfs-zed
+      modprobe -r zfs
+      modprobe zfs
+      systemctl start zfs-zed
+    fi
   fi
 }
 
@@ -568,7 +570,12 @@ function custom_install_operating_system {
 function install_zfs_0.8_packages {
   print_step_info_header
 
-  chroot_execute "add-apt-repository --yes ppa:jonathonf/zfs"
+  if [[ "${c_linux_setups_zfs_0_8_support["$v_linux_distribution"]}" == "1" ]]; then
+    chroot_execute "add-apt-repository --yes ppa:jonathonf/zfs"
+  else
+    chroot_execute "apt update"
+  fi
+
   chroot_execute 'echo "zfs-dkms zfs-dkms/note-incompatible-licenses note true" | debconf-set-selections'
   chroot_execute "apt install --yes zfs-initramfs zfs-dkms grub-efi-amd64-signed shim-signed"
 }
